Update raw401.hushNickOfflineHandler.kvs
20191230

1. Ensure G_hushLive{"iTransfersActive"} value doesn't go under value of 0.

Notes: when a bot is offline from IRC and you request a pack from it, the IRC server replies with a "no such nick" message. The raw 401 handler deals with this. If you have "Under the Radar" setting ticked, you won't see any feedback about Hush handling this situationn unless the handling of it fails in some way.